Refactor the handlers module to have it's own folder
We are refactoring the DNF and YUM modules to be in a separate folder,
as, we are also placing the callback classes in it's own module, mostly
because we are placing as well the licenses of the modules that we are
using and adaptating from.

Since we had to separate the callback classes to a new module and not
place them in the previous `dnf.py` and `yum.py`, we went ahead and
moved also the `dnf.py` and `yum.py` to be an `__init__.py`, but they
are placed under their respective folders.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>